fix: 도메인에 의존하도록 각 반환 타입을 UserEntity -> User로 변경
