--- src/main/java/com/iroff/supportlab/presentation/config/global/security/CustomUserDetailsService.java
@@ -1,24 +1,0 @@
-package com.iroff.supportlab.presentation.config.global.security;
-
-import com.iroff.supportlab.presentation.user.out.persistence.UserEntity;
-import com.iroff.supportlab.presentation.user.out.persistence.UserJpaRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
-
-    private final UserJpaRepository userJpaRepository;
-
-    // todo: 사용자 지정 Exception 추가 시 수정 필요
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        UserEntity user = userJpaRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("해당 사용자가 없습니다."));
-        return new CustomUserDetails(user);
-    }
-}
